feat: rename base package
